refactor(android): extract shared JNI helpers into JniSupport
ScopedJNIEnv, throwIfJavaExceptionPending and a new resolveGlobalClass
helper move out of platform_android.cpp's local namespace so the
upcoming HTTP JNI bridge can reuse them instead of duplicating the
attach/detach and class-caching logic. No behavior change — verified
with a real CMake/NDK build of the example app.
